Type fixups for libc::strlen()
The libc version of `strlen()` takes a `const *i8` and returns a `usize`,
which is a bit different from `str_len()` which took a `const *u8` and
returns a `u32`.

This fixes all occurrances so (hopefully) the types match up!

+++ src/cachetest.rs
@@ -76,7 +76,7 @@ pub unsafe extern "C" fn _c_main(mut argc: i32, mut argv: *mut *mut u8) -> i32 {
                 x as (*const u8),
                 i as (u32),
                 x.offset(i as (isize)).offset(1isize) as (*const u8),
-                libc::strlen(x as (*const u8))
+                libc::strlen(x as *const i8) as u32
                     .wrapping_sub(i as (u32))
                     .wrapping_sub(1u32),
                 86400u32,
